Add axios

+++ src/main/java/langolier/portfolio/repos/ProjectRepo.java
@@ -3,6 +3,6 @@ package langolier.portfolio.repos;
 import langolier.portfolio.models.Project;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjectRepo extends CrudRepository<Project, Integer> {
+public interface ProjectRepo extends CrudRepository<Project, Long> {
 
 }
